examples/crewai/scifi_writer/requirements.txt
python-dotenv
crewai
